+++ src/frontend/history.html
@@ -1,45 +1,35 @@
-import pymongo
-import webbrowser
-
-client = pymongo.MongoClient("mongodb://localhost:27017/")
-mydb = client["myDB"]
-mycol = mydb["builds"]
-
-student = []
-
-tbl = "<tr><td>Name</td><td>Age</td><td>gender</td></tr>"
-student.append(tbl)
-
-for y in mycol.find():
-    a = "<tr><td>%s</td>"%y['name']
-    student.append(a)
-    b = "<td>%s</td>"%y['age']
-    student.append(b)
-    c = "<td>%s</td></tr>"%y['gender']
-    student.append(c)
-
-contents = '''<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-<html>
+<html lang="en">
 <head>
-<meta content="text/html; charset=ISO-8859-1"
-http-equiv="content-type">
-<title>History</title>
-<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
+    <title>History</title>
 </head>
 <body class="bg-dark text-light">
-<table CLASS="table">
-%s
-</table>
+    
+    <div class="container-fluid">
+        <div class="row jumbotron text-center mt-3">
+            <h2 class="display-5">CI Server Documentation</h2>
+        </div>
+        <nav class="navbar justify-content-center">
+            <ul class="nav">
+                <li class="nav-item">
+                  <a class="nav-link" href="server.html">Server</a>
+                </li>
+                <li class="nav-item">
+                  <a class="nav-link" href="notifications.html">Notifications</a>
+                </li>
+                <li class="nav-item">
+                  <a class="nav-link" href="tests.html">Tests</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="history.html">History</a>
+                </li>
+            </ul>
+        </nav>
+        {{replace}}
+    </div>
+    
 </body>
 </html>
-'''%(student)
-
-filename = 'info.html'
-
-def main(contents, filename):
-    output = open(filename,"w")
-    output.write(contents)
-    output.close()
-
-main(contents, filename)    
-webbrowser.open(filename)
