+++ src/frontend/index.html
@@ -19,7 +19,6 @@
     <div class="container-fluid">
         <div class="row jumbotron text-center mt-3">
             <h2 class="display-5">CI Server Documentation</h2>
-            <!-- <p class="lead">Description</p> -->
         </div>
 
         <nav class="navbar justify-content-center">
