+++ src/frontend/server.html
@@ -1,0 +1,36 @@
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-gH2yIJqKdNHPEq0n4Mqa/HGKIhSkIHeL5AyhkYV8i59U5AR6csBvApHHNl/vI1Bx" crossorigin="anonymous">
+    <title>Server</title>
+</head>
+<body class="bg-dark text-light">
+    
+    <div class="container-fluid">
+        <div class="row jumbotron text-center mt-3">
+            <h2 class="display-5">CI Server Documentation</h2>
+        </div>
+        <nav class="navbar justify-content-center">
+            <ul class="nav">
+                <li class="nav-item">
+                  <a class="nav-link" href="server.html">Server</a>
+                </li>
+                <li class="nav-item">
+                  <a class="nav-link" href="notifications.html">Notifications</a>
+                </li>
+                <li class="nav-item">
+                  <a class="nav-link" href="tests.html">Tests</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="history.html">History</a>
+                </li>
+            </ul>
+        </nav>
+
+        
+    </div>
+    
+</body>
+</html>
